+++ src/main/java/com/acelerati/stock/domain/model/model/product/Product.java
@@ -1,4 +1,4 @@
-package com.acelerati.stock.domain.model.product;
+package com.acelerati.stock.domain.model.model.product;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,8 +7,11 @@ import lombok.Setter;
 @AllArgsConstructor
 @Getter
 @Setter
-public class Brand {
+public class Product {
     private Long id;
     private String name;
+    private String model;
+    private Brand brand;
+    private Category category;
     private String description;
 }
